+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Mapping for parsing `MARC` records
 - Manager for generating PDF manifest files (in JSON-LD)
 - Process for ingesting records from Project MUSE
+### Fixed
+- Fix timeout issues with RabbitMQ during long running processes
+- Handle odd record formatting in NYPL MARC records
 
 ## 2021-01-06 -- v0.0.5
 ### Added